sua giao dien+image+inbound button

+++ MobilePhoneManagementSystem/src/outbound/controller/OutboundDialog.java
@@ -672,6 +672,9 @@ public class OutboundDialog extends javax.swing.JDialog implements ItemListener 
         jScrollPane3.setViewportView(list);
 
         jButton2.setIcon(new javax.swing.ImageIcon(getClass().getResource("/image/product/refresh_1.png"))); // NOI18N
+        jButton2.setBorderPainted(false);
+        jButton2.setContentAreaFilled(false);
+        jButton2.setFocusPainted(false);
         jButton2.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
                 jButton2ActionPerformed(evt);
@@ -721,10 +724,11 @@ public class OutboundDialog extends javax.swing.JDialog implements ItemListener 
                 .addGap(2, 2, 2)
                 .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(btnSave)
-                    .addComponent(jButton5)
-                    .addComponent(btReset, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(btReset, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                        .addComponent(btnSave)
+                        .addComponent(jButton5)))
                 .addGap(10, 10, 10))
         );
 
@@ -774,6 +778,8 @@ public class OutboundDialog extends javax.swing.JDialog implements ItemListener 
     }//GEN-LAST:event_jButton5ActionPerformed
 
     private void btDeleteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btDeleteActionPerformed
+        int ans = SwingUtils.showConfirmDialog("Are you sure to delete? Quantity of such product will add back to stock");
+        if(ans==JOptionPane.YES_OPTION)
         deleteAction();
     }//GEN-LAST:event_btDeleteActionPerformed
